+++ lib/models/money_objects/account_aliases/account_alias.dart
@@ -44,6 +44,8 @@ class AccountAlias extends MoneyObject {
 
   @override
   int get uniqueId => id.value;
+  @override
+  set uniqueId(value) => id.value = value;
 
   /// Constructor from a SQLite row
   @override