+++ lib/models/money_objects/loan_payments/loan_payment.dart
@@ -7,6 +7,8 @@ import 'package:money/widgets/list_view/list_item_card.dart';
 class LoanPayment extends MoneyObject {
   @override
   int get uniqueId => id.value;
+  @override
+  set uniqueId(value) => id.value = value;
 
   /// ID
   /// 0|Id|INT|1||0