+++ lib/models/money_objects/money_objects.dart
@@ -1,10 +1,11 @@
+// Imports
 import 'package:flutter/foundation.dart';
 import 'package:money/models/settings.dart';
 import 'package:money/storage/data/data.dart';
 import 'package:money/storage/database/database.dart';
 import 'package:money/models/money_objects/money_object.dart';
 
-// exports
+// Exports
 export 'package:money/models/money_objects/money_object.dart';
 export 'package:money/models/fields/fields.dart';
 export 'package:collection/collection.dart';
@@ -55,6 +56,10 @@ class MoneyObjects<T> {
 
     // keep track of new items, they will need to be persisted later
     if (isNewEntry) {
+      if (entry.uniqueId == -1) {
+        entry.uniqueId = length + 1;
+      }
+
       entry.mutation = MutationType.inserted;
       Data().notifyTransactionChange(MutationType.inserted, entry);
     }
@@ -99,13 +104,13 @@ class MoneyObjects<T> {
         case MutationType.none:
           break;
         case MutationType.inserted:
-          db.insert(tableName, item.getPersistableJSon());
+          db.insert(tableName, item.getPersistableJSon<T>());
 
         case MutationType.deleted:
           db.delete(tableName, item.uniqueId);
 
         case MutationType.changed:
-          db.update(tableName, item.uniqueId, item.getPersistableJSon());
+          db.update(tableName, item.uniqueId, item.getPersistableJSon<T>());
 
         default:
           debugPrint('Unhandled change ${item.mutation}');