+++ lib/models/money_objects/online_accounts/online_account.dart
@@ -28,6 +28,8 @@ import 'package:money/models/money_objects/money_objects.dart';
 class OnlineAccount extends MoneyObject {
   @override
   int get uniqueId => id.value;
+  @override
+  set uniqueId(value) => id.value = value;
 
   // 0
   Field<OnlineAccount, int> id = Field<OnlineAccount, int>(